requirements.txt
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.1
cffi==1.14.0
cryptography==2.8
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.0.2
django-appconf==1.0.3
django-autocomplete-light==3.5.1
django-bootstrap-modal-forms==2.0.1
django-bootstrap-pagination==1.7.1
django-braces==1.14.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-cryptography==1.0
django-discover-runner==1.0
django-extensions==3.0.9
django-rest-auth==0.9.5
django-restframework==0.0.1
django-tempus-dominus==5.1.2.13
django-weasyprint==1.0.2
django-widget-tweaks==1.4.5
django-wkhtmltopdf==3.3.0
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
#gunicorn==20.0.4
html5lib==1.0.1
install==1.3.1
lxml==4.6.2
Pillow==7.1.2
pycparser==2.19
pycrypto==2.6.1
pylokit==0.8.1
PyPDF2==1.26.0
Pyphen==0.10.0
python-docx==0.8.10
pytz==2019.3
reportlab==3.5.59
six==1.14.0
sqlparse==0.3.0
style==1.1.0
tinycss2==1.1.0
update==0.0.1
WeasyPrint==52.2
webencodings==0.5.1
xhtml2pdf==0.2.4
